Split CEN into 1B274 and D600
