+++ config/splat.us.stcen.yaml
@@ -31,6 +31,7 @@ segments:
       - [0x1300, data]
       - [0xD42C, rodata]
       - [0xD600, c]
+      - [0x1B274, c]
       - [0x1C704, rodata]
       - [0x1C764, rodata]
       - [0x1D374, rodata]
