+++ src/st/cen/cen.h
@@ -25,7 +25,7 @@ typedef enum {
 } EntityIDs;
 
 extern PfnEntityUpdate D_80180390[];
-extern u16 D_80180410[];
+extern u16 D_80180410[]; // Init
 extern u16 D_80180428[];
 extern u16 D_80180458[];
 extern ObjInit2 D_80180490[];
@@ -33,6 +33,9 @@ extern u16 D_801805D8[];
 extern s32 D_8018068C[];
 extern s16 D_80180BBC[];
 extern s8 D_80180FE8[];
+extern u16 D_801811C8[];
+extern u16 D_801811D8[];
+extern u8 D_80181238;
 extern ObjInit2 D_8018125C[];
 void CreateEntityFromCurrentEntity(u16 objectId, Entity* entity);
 extern LayoutObject* D_8019C764;
@@ -50,6 +53,7 @@ extern s16 D_8019D38A;
 extern s8 D_8019D38E;
 extern s8 D_8019D38F;
 void func_80192A3C(Entity*, LayoutObject*);
+extern u16 func_80194E44(u16 arg0, u16 arg1, u16 arg2);
 extern u8 D_8019C76C;
 extern u8 D_8019C770;
 #endif
